requirements.txt
langchain==0.3.25
google-genai==1.14.0
faiss-cpu==1.11.0
fastapi==0.115.9
pymssql==2.3.4
transformers==4.51.3
sentence-transformers==4.1.0
numpy==2.2.5
pandas==2.2.3
pydantic==2.11.4
uvicorn==0.34.2
langchain-community==0.3.24